project-1/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
idna==3.3
imageio==2.14.1
imutils==0.5.4
networkx==2.6.3
numpy==1.22.1
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.0
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
requests==2.27.1
scikit-image==0.19.1
scipy==1.7.3
six==1.16.0
tifffile==2021.11.2
urllib3==1.26.8